Move VarExporter to support
