+++ src/Viserio/Component/Support/Tests/Fixture/Controller.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+namespace Viserio\Component\Support\Tests\Fixture;
+
+class Controller
+{
+    public function string()
+    {
+        return 'test';
+    }
+}
