+++ src/Viserio/Component/Support/VarExporter.php
@@ -1,6 +1,6 @@
 <?php
 declare(strict_types=1);
-namespace Viserio\Component\Routing;
+namespace Viserio\Component\Support;
 
 use stdClass;
